Limit versions of ansible-core we use to include necessary fix
Upstream in ansible-core devel the connection plugin backend fixed a bug
that was previously requiring us to set self._sub_plugin in the logix
connection plugin. However, now that the fix is in we must NOT set
self._sub_plugin or core will traceback at the task executor.

From now on we will limit via meta/runtime.yml for the collection to
require ansible-core >=2.12.9 for the 2.12.x series, >=2.13.4 for the
2.13.x series, and anything newer than that (the fix was initially
released in 2.14.0 and backported to the other stable releases).

PR#74446 (commit id bf1ef5a1f3562c9a59168adbc78750304c3e4309)
https://github.com/ansible/ansible/pull/74446

Also included in this PR is various ansible-lint cleanup.

Signed-off-by: Adam Miller <admiller@redhat.com>